+++ packages/react/src/lib/accessibility/Behaviors/Grid/gridHorizontalBehavior.ts
@@ -3,7 +3,7 @@ import { FocusZoneDirection } from '../../FocusZone'
 
 /**
  * @description
- * Provides navigation between focusable children of Grid component with 4 arrow keys in horizontal direction (based on DOM order).
+ * Provides navigation between focusable children of Grid component with arrow keys in horizontal direction (based on DOM order).
  * Right/Down arrow keys move to next item, Up/Left  arrow keys to previous item. Right and Left arrow keys are switched in RTL mode.
  *
  * @specification